+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 
-## [0.7] - Unreleased
+## [0.7] - 2024-05-23
 
 - Fix bug in `e` keybinding (`pgmacs-run-sql`) when outside a table.